+++ polaris_sql/migrations/2026-09-25-005-retention-set-through-a-procedure.down.sql
@@ -1,0 +1,2 @@
+-- Reverts 2026-09-25-005.
+DROP PROCEDURE IF EXISTS uc_set_retention_policy(varchar, varchar, integer, text, integer, integer, integer);